Moved dev stuff to actual stuff
